updates model generation logic
This change updates the algorithms used to generate the models and adds
command line options which expose various parameters that can be tweaked
for model training and interaction. The model outputs can be selected from
the command line. The epoch can be set from the command line. The filename
to save the generated model can be selected from the command line.

An action option allows the user to select between model training and
model interaction. When interacting with the model, the model to load
can be selected and the input values can be supplied through command
line options.